// packages/dnb-eufemia/src/extensions/forms/types.d.ts
export type AllJSONSchemaVersionsBasis<DataType> = JSONSchema4 | JSONSchema6 | JSONSchema7 | JSONSchemaType<DataType>;
/**
 * Union type supporting all JSON Schema (AJV).
 * IMPORTANT: When using these types, you MUST provide ajvInstance to Form.Handler
 */
export type AllJSONSchemaVersions<DataType = unknown> = AllJSONSchemaVersionsBasis<DataType> | (Omit<AllJSONSchemaVersionsBasis<DataType>, 'required'> & {
    required?: readonly string[];
});
export type { JSONSchemaType };
export type Schema<Value = unknown> = AllJSONSchemaVersions<Value> | ZodSchema;
export type ValidatorReturnSync<Value> = Error | FormError | string | React.ReactElement | undefined | void | Array<Validator<Value>> | Array<Error | FormError>;
export type ValidatorReturnAsync<Value> = ValidatorReturnSync<Value> | Promise<ValidatorReturnSync<Value>>;
export type Validator<Value, ErrorMessages = DefaultErrorMessages> = (value: Value, additionalArgs: ReceiveAdditionalEventArgs<Value, ErrorMessages>) => ValidatorReturnAsync<Value>;
// ...
export type DotNotationErrorMessages = Record<`${string}` | `${string}.${string}`, string | React.ReactElement>;
/**
 * { 'nb-NO': { 'Field.errorRequired': 'Dette feltet er påkrevd' } }
 */
export type ErrorMessagesWithLocaleSupport = Record<FormsTranslationLocale, DefaultErrorMessages>;
export type InternalErrorMessages = Record<Exclude<FormsTranslationFlat, undefined>, string>;
export type DefaultErrorMessages = Partial<InternalErrorMessages> & Partial<DotNotationErrorMessages>;
// ...
export type ProvideAdditionalEventArgs = Record<string, unknown>;
export type DataValueReadComponentProps<Value = unknown> = ComponentProps & DataValueReadProps<Value>;
export type DataValueReadWriteComponentProps<Value = unknown, EmptyValue = undefined | unknown> = ComponentProps & DataValueReadProps<Value> & DataValueWriteProps<Value, EmptyValue>;
// ...
export type FieldStatus = {
    info?: StatusMessage;
    warning?: StatusMessage;
    error?: StatusError;
};
export type ConnectorProps<Value = unknown> = Pick<UseFieldProps<Value>, 'onChange' | 'onBlurValidator'>;
export type InfoProp<Value> = MessageProp<Value, React.ReactNode | Array<React.ReactNode>>;
export type WarningProp<Value> = MessageProp<Value, React.ReactNode | Array<React.ReactNode>>;
export type ErrorProp<Value> = MessageProp<Value, string | React.ReactElement | Error | FormError | Array<string | React.ReactElement | Error | FormError>>;
export interface UseFieldProps<Value = unknown, EmptyValue = undefined | unknown, ErrorMessages extends DefaultErrorMessages = DefaultErrorMessages, ExtraValue extends ProvideAdditionalEventArgs = ProvideAdditionalEventArgs> extends DataValueReadWriteComponentProps<Value, EmptyValue>, AriaAttributes {
    /**
     * ID added to the actual field component, and linked to the label via for-attribute
     */
    id?: Identifier;
    name?: string;
    disabled?: boolean;
    readOnly?: boolean;
    autoComplete?: HTMLInputElement['autocomplete'] | HTMLTextAreaElement['autocomplete'];
    /**
     * Text showing in place of the value if no value is given
     */
    placeholder?: React.ReactNode;
    /**
     * NB: Undocumented for now.
     * Forwards all possible props to the underlying component.
     */
    htmlAttributes?: Record<string, unknown>;
    /**
     * NB: Undocumented for now.
     * Forwards all given props in a props object.
     */
    props?: Record<string, unknown>;
    info?: InfoProp<Value>;
    warning?: WarningProp<Value>;
    error?: ErrorProp<Value>;
    /**
     * Called when the current status (info, warning or error) changes.
     * Provides the latest values for each status key.
     */
    onStatusChange?: (status: FieldStatus) => void;
    required?: boolean;
    schema?: Schema<Value> | ((props: UseFieldProps<Value>) => Schema<Value>);
    onChangeValidator?: Validator<Value>;
    onBlurValidator?: Validator<Value>;
    exportValidators?: Record<string, Validator<Value>>;
    validateRequired?: (internal: Value, { emptyValue, required, isChanged, error, }: {
        emptyValue: EmptyValue;
        required: boolean;
        isChanged: boolean;
        error: FormError | undefined;
    }) => FormError | undefined;
    /**
     * Should error messages based on validation be shown initially (from given value-prop or source data)
     * before the user interacts with the field?
     * @default false
     */
    validateInitially?: boolean;
    /**
     * Should error messages be shown when touching (like focusing a field and blurring) without having changed
     * the value? So the user did not introduce a new error, but it was invalid based on validation initially.
     */
    validateUnchanged?: boolean;
    /**
     * Should validation be done while writing, not just when blurring the field?
     */
    validateContinuously?: boolean;
    /**
     * Provide custom error messages for the field
     */
    errorMessages?: ErrorMessages;
    /**
     * Transforms the `value` before it's displayed in the field (e.g. input).
     * Public API. Should not be used internally.
     */
    transformIn?: (external: unknown) => Value | ExtraValue;
    /**
     * Transforms the value before it gets forwarded to the form data object or returned as the onChange value parameter.
     * Public API. Should not be used internally.
     */
    transformOut?: (internal: Value | ExtraValue, additionalArgs?: unknown) => unknown;
    /**
     * Transforms the value given by `handleChange` after `fromInput` and before `updateValue` and `toEvent`. The second parameter returns the current value.
     */
    transformValue?: (value: Value, currentValue?: Value) => Value;
    /**
     * Transform additionalArgs or generate it based on `value`.
     */
    provideAdditionalArgs?: (value: Value, additionalArgs?: ProvideAdditionalEventArgs) => ProvideAdditionalEventArgs;
    /**
     * Transforms the value before it gets returned as the `value`.
     */
    toInput?: (external: Value | unknown) => Value | unknown;
    /**
     * Transforms the internal value before it gets returned by even callbacks such as `onChange`, `onFocus` and `onBlur`. The second parameter returns the event type: `onChange`, `onFocus`, `onBlur` or `onBlurValidator`.
     */
    toEvent?: (internal: Value, type: 'onChange' | 'onFocus' | 'onBlur' | 'onBlurValidator') => Value;
    /**
     * Transforms the value given by `handleChange` before it is used in the further process flow. Use it to destruct the value from the original event object.
     */
    fromInput?: (external: Value | unknown) => Value;
    /**
     * Transforms the given props `value` before any other step gets entered.
     */
    fromExternal?: (external: Value) => Value;
    /**
     * For internal use only.
     */
    valueType?: string | number | boolean | Array<string | number | boolean>;
}
export type FieldProps<Value = unknown, EmptyValue = undefined | unknown, ErrorMessages extends DefaultErrorMessages = DefaultErrorMessages, ExtraValue extends ProvideAdditionalEventArgs = ProvideAdditionalEventArgs> = UseFieldProps<Value, EmptyValue, ErrorMessages, ExtraValue> & SharedFieldBlockProps;
// ...
export type Identifier = string;